--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@datnguyen1215/hsmjs",
-  "version": "1.1.2",
+  "version": "1.1.3",
   "description": "A lightweight, imperative state machine library with hierarchical states and async action support",
   "main": "dist/cjs/index.js",
   "module": "dist/es/index.js",